BlogAPI_DjangoRF/requirements.txt
asgiref==3.2.5
Django==3.0.4
django-crispy-forms==1.9.0
django-filter==2.2.0
django-markdown-deux==1.0.5
django-pagedown==2.0.3
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
Markdown==3.2.1
markdown2==2.3.8
Pillow==7.0.0
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.1
